watercostaccra1: Household survey data

A survey of 116 households was conducted in two low-income communities in metropolitan Accra. These are Korle Gonno, a larger, well-planned coastal area with over 35 household water vendors, and Abuja, a small, densely packed, extralegal settlement with 15 water vendor and bathhouse businesses.

Usage

watercostaccra1

Format

A tibble with 116 rows and 40 variables

id

identification number of household

community

one of two communities surveyed (Korle Gonno or Abuja)

housing_type

housing type (1 block unit: unit in a row of apartments made of cement blocks, 2 wood unit: unit in a row of apartments made of wood, house, 3 compound house: single-story L- or C-shaped house with a multiple units around a shared courtyard, 4 multi-story apartment building, 5 wooden shack, 6 no structure, 7 other)

respondent_relationship_to_hh

respondent's relationship to the household head (respondent identified)

gender

gender of respondent (respondent identified)

tenure

tenure status (renter, homeowner, or living without payment)

years_in_community

number of years respondent has lived in community

adult_count

number of adults in household including respondent. Household is described as those "eating from the same pot"

child_count

number of children under 18 in household. Household is described as those "eating from the same pot"

rooms_in_hh

number of rooms used for sleeping. Household is described as those "eating from the same pot"

business_ownership

household or respondent owns a business (respondent-owned or household-owned)

business_location

home-based, fixed location outside home, or mobile location

business_category

type of business (e.g., salon, shop, water vending)

business_water_use

respondent's business uses water beyond typical needs of household (true or false)

business_water_source

primary source of water for business use (packaged water, piped to home, piped to neighbor's home, piped to compound, commercial or public tap, borehole, dug well, spring water, delivered water)

primary_dw_source

primary source of drinking water (packaged water, piped to home, piped to neighbor's home, piped to compound, commercial or public tap, borehole, dug well, spring water, delivered water)

dw_reason_x

respondent reasons for using drinking water source (convenience, affordability, availability, temperature, cleanliness, taste, habit or cultural norm, trustworthiness, health, other)

package_type_preference

respondent typically purchases individual sachets/bottles, multipacks of these, or both

package_size_reason_x

reason for purchasing preferred package type (storage space in home, cost effectiveness, temperature at time of purchase, availability of money, convenience, size needed for respondent or household, avoiding wasting water by purchasing when needed)

dw_treatment

treatment methods of water before drinking

primary_water_source

primary water source for non-drinking water (packaged water, piped to home, piped to neighbor's home, piped to compound, commercial or public tap, borehole, dug well, spring water, delivered water)

primary_source_reason_x

reason for using primary source of non-drinking water (proximity to home, convenience, affordability, availability, cleanliness, other)

other_non_dw_source_use

respondent uses at least one source besides primary non-drinking water source (true or false)

other_non_dw_sources_x

additional water source(s) for non-drinking water (packaged water, piped to home, piped to neighbor's home, piped to compound, commercial or public tap, borehole, dug well, spring water, delivered water)

secondary_source_reason_x

reason for using secondary source of non-drinking water (primary source is not available, primary source is not clean, primary source is crowded, availability of shower stalls, convenient location)

tap_payment_mode

respondent's mechanism for paying for piped water (all respondents use piped water as a primary or secondary source), options including pay_to_fetch, shares_bill, and both.

daily_hh_water_cost_for_pay_to_fetch

daily estimated cost of drinking water for respondent's household

daily_hh_water_cost_phhm_for_pay_to_fetch

daily estimated cost of drinking water for respondent's household per household member

past_struggle_to_find_water

respondent has struggled to find water before (defined as extreme difficulty to access water) (true or false)

time_of_last_struggle_to_find_water

respondent's last time of struggle to find water (e.g., in the last week)

weekdays_struggle_to_find_water

days in a week the respondent typically struggles to find or pay for water

past_struggle_primary_reason

primary reason for past struggles to find water (availability, high cost, distance to nearest source)

tap_closure_knowledge_x

respondent's knowledge about tap closures (usually known, sometimes known, expected due to patterns in closures, not known, or no answer)

coping_mechanism_x

strategies for coping with water shortage (spending more on the same amount of water, purchasing extra water to store at home, using another source, using packaged water for cooking, skipping cooking, using packaged water for bathing, skipping bathing, closing business due to water shortage, skipping laundry)

water_storage_drinking_water

respondent typically stores drinking water at home (true or false)

water_storage_non_drinking_water

respondent typically stores non-drinking water at home (true or false)

water_storage_none

respondent typically does not store water at home (true or false)

storage_containers_x

if respondent typically stores non-drinking water, types of storage containers (plastic jugs also called jerry cans or Kufuor gallons, uncovered or covered barrels, other covered or uncovered containers)

estimated_non_dw_storage_capacity

estimated capacity of storage for non-drinking water in liters

estimated_stored_non_dw

estimated actual stored non-drinking water in liters