watercostaccra2: Water point survey data

A water point survey was conducted in two low-income communities in metropolitan Accra. These are Korle Gonno, a larger, well-planned coastal area with over 35 household water vendors, and Abuja, a small, densely packed, extralegal settlement with 15 water vendor and bathhouse businesses.

Usage

watercostaccra2

Format

A tibble with 49 rows and 28 variables

id

identification number

community

one of two communities surveyed (Korle Gonno or Abuja)

type

water point type (piped water, borehole, dug well, natural spring)

available_services

services available at water point (bathing, public sale of water, toilet)

location

within a compound or outside compound on a street

year_established

year established

owner

owner, options including 1 household head, 2 household member(s), 3 community member outside household, and 4 multiple community members outside household

constructor

constructor (government or community member)

managers

typical manager(s) of water point (household head or member(s), employee(s), self managed by customers)

estimated_storage_capacity_liters

estimated storage capacity in liters

average_visits_per_customer

average number of daily visits per customer

respondent_would_use_to_prepare_rice

respondent would use this water to prepare rice (true or false)

perception_of_quality

respondent's perception of water quality

tap_closure_days_per_week

typical number of tap closures per week

price_25_liter_jug

current price of 25-liter jug of water (cedis)

price_20_liter_bucket

current price of 20-liter bucket of water (cedis)

price_30_liter_basin

current price of 30-liter basin of water (cedis)

avg_price_per_liter_cedis

average price per liter, calculated by averaging price per liter of known prices (cedis)

tap_closure_changes

typical dynamics of water point management during closure (increasing prices, water point likely to close due to low storage, bathing customers have less water than when taps are flowing)

flexible_pricing

manager adjusts price depending on amount of water needed or familiarity or need of customer (true or false)

price_increase

price of any volume of water has increased in the last year (true or false)

CBT_sample_source

source of sample for compartment bag test (CBT) supplied by Aquagenx (https://www.aquagenx.com/cbt-ectc/)

coli_mpn

results of E. Coli most probable number (MPN) test per 100 mL sample

coli_mpn_ci

results of E. Coli most probable number (MPN) test per 100 mL sample - upper 95% confidence interval (CI)

coli_mpn_health_risk

results of E. Coli most probable number (MPN) test per 100 mL sample - descriptive health risk

tc_mpn

results of Total Coliforms (TC) most probable number (MPN) test per 100 mL sample

tc_mpn_ci

results of Total Coliforms (TC) most probable number (MPN) test per 100 mL sample - upper 95% confidence interval (CI)

tc_mpn_health_risk

results of Total Coliforms (TC) most probable number (MPN) test per 100 mL sample - descriptive health risk